Move TwinkleConfig closer to parent config
